Factor out BoardScatterPlane
